feat: add isKnownError/isValidationError guards, move errors to top level
Move the errors module from src/lib/errors.js to src/errors.js (it's a
public entry point, not an internal lib) and update all importers.

Add two typeguards, both narrowing `err.code` to the matching union of
codes:
- `isKnownError(err)` — the error was thrown by this module.
- `isValidationError(err)` — an expected validation error (bad input or
  invalid file): every known error except the programmer-error
  `AddAfterFinishError`. Lets consumers show a friendly message and skip
  error reporting (e.g. Sentry).

A single KNOWN_ERRORS array drives both the runtime check and the
KnownError / KnownErrorCode / ValidationError types, so adding an error
keeps both in sync. Document the new exports in the README.
