+++ bin/helpers/read-files.js
@@ -4,6 +4,7 @@ import path from 'node:path'
 import parseJson from 'parse-json'
 import * as v from 'valibot'
 
+import { SchemaError } from '../../src/errors.js'
 import {
 	CATEGORIES_DIR,
 	FIELDS_DIR,
@@ -12,7 +13,6 @@ import {
 	METATADATA_FILE,
 	PRESETS_DIR,
 } from '../../src/lib/constants.js'
-import { SchemaError } from '../../src/lib/errors.js'
 import { parseSvg } from '../../src/lib/parse-svg.js'
 import { isNotFoundError } from '../../src/lib/utils.js'
 import {
